# Upgrade babel by switching to swc